Bugfixes in `defaultTemplateOf`, comment improv.
Hate untypes languages. Fixed two bugs:

Bug 1: When refactoring the last time, I happened to leave `topDir` variable in
`find`, but it's not needed there, it's needed in `defualtTemplateOf`, the
function I extracted.

Bug 2: I tested what happened if someone deleted the default template file, and
the code we wrote did climb out of the toplevel directory. Reason? When
comparing, we did not add a slash at the end.