Added tests, discovered 2 bugs, fixed 1.
Added some small tests to find-template, these can problably be improved
and expanded in the future. I don't even know how you're "supposed" to do
tests in nodejs.

Anyways, these tests made me discover the bug that, if `defualtTemplateOf`
gets a directory that is above the template dir, it will go into an
infinite loop. So I added a safeguard there.

I also discovered a thing in routes: if you have no internet, the request
to taitan will throw an error that is not caught with the error handler.
Look like it's throwing and error event with it's event handler. Not sure
how to catch that, but I'll be submitting an issue.