Refactored config to focus on subdomains
Hope everything works well, ran the tests and fixed issues. Now every
subdomain in use should be declared in "subdomainData" and we should use
settings according to that description. We should in the future not guess
the supported engine at all. But, we must add behavior for when the right
engine is not found.